Moved _has_tag function to utils.__init__
