+++ sktime/utils/__init__.py
@@ -8,6 +8,8 @@ from importlib import import_module
 from operator import itemgetter
 from pathlib import Path
 
+from sktime.tests._config import VALID_ESTIMATOR_TAGS
+
 
 def _get_name(obj):
     return obj.__class__.__name__
@@ -176,3 +178,21 @@ def all_estimators(estimator_types=None, return_names=True, exclude_estimators=N
         return all_estimators
     else:
         return [estimator for (name, estimator) in all_estimators]
+
+
+def _has_tag(Estimator, tag):
+    """Check whether an Estimator has the given tag or not.
+
+    Parameters
+    ----------
+    Estimator : Estimator class
+    tag : str
+        An Estimator tag like "skip-inverse-transform"
+
+    Returns
+    -------
+    bool
+    """
+    assert tag in VALID_ESTIMATOR_TAGS
+    # Check if tag is in all tags
+    return Estimator._all_tags().get(tag, False)
